move Category component into the containers folder
