+++ src/containers/App/App.tsx
@@ -1,7 +1,7 @@
 import Header from "../../components/Header";
 import { Route, Switch } from "react-router-dom";
 import Categories from "../../components/Categories";
-import Category from "../../components/Category";
+import Category from "../Category";
 import Item from "../../components/Item";
 import Cart from "../../components/Cart";
 import { StyledApp } from "../../components/styles/App.styled";
